Fix code tests
