+++ modules/entities/index.js
@@ -1,2 +1,5 @@
+export { entityChangeset } from './changeset';
 export { entityEntity } from './entity';
 export { entityNode } from './node';
+export { entityRelation } from './relation';
+export { entityWay } from './way';
