+++ modules/index.js
@@ -37,8 +37,8 @@ export { geoVecFloor as geoRoundCoordinates } from './geo/vector';
 export { geoVecLength as geoEuclideanDistance } from './geo/vector';
 export { entityEntity as Entity } from './entities/entity';
 export { entityNode as Node } from './entities/node';
-export { osmRelation as Relation } from './osm/relation';
-export { osmWay as Way } from './osm/way';
+export { entityRelation as Relation } from './entities/relation';
+export { entityWay as Way } from './entities/way';
 export { rendererBackgroundSource as BackgroundSource } from './renderer/background_source';
 export { rendererBackground as Background } from './renderer/background';
 export { rendererFeatures as Features } from './renderer/features';
