+++ test/index.html
@@ -78,6 +78,8 @@
     <script src='spec/core/tree.js'></script>
     <script src='spec/core/validator.js'></script>
 
+    <script src='spec/entities/entity.js'></script>
+
     <script src='spec/geo/extent.js'></script>
     <script src='spec/geo/geo.js'></script>
     <script src='spec/geo/geom.js'></script>
